requirements.txt
asgiref==3.8.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
Django==5.0.4
djangorestframework==3.15.1
Flask==3.0.3
flask-swagger-ui==4.11.1
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
sqlparse==0.5.0
tzdata==2024.1
Werkzeug==3.0.2
